Added parser, changed scanner language
Changed scanner language to NOT include symbols found in the input language.

Added a counter to the scanner to make sure all parens have a match.

Added a parser. Parser currenty accepts silently, but has places to pass symbols to other methods to build a state diagram.

Added invalid_next() as a helper to check that a (, |, *, or + is not followed by an invalid symbol.